+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Improved keyboard navigation
 * Added image scaling to images, so that images always fit the window
+* Added search bar to look people up
 
 ## Technical